backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-bigquery==3.23.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1